+++ src/main/java/raisetech/studentManagement/StudentManagementApplication.java
@@ -6,25 +6,14 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.studentManagement.data.Student;
+import raisetech.studentManagement.data.StudentsCourses;
 
 @SpringBootApplication
 @RestController
 public class StudentManagementApplication {
 
-	@Autowired
-	private StudentRepository repository;
-
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
-@GetMapping("/studentList")
-	public List<Student> getStudentList() {
-		return repository.search();
-	}
-
-@GetMapping("/studentCourseList")
-public List<StudentsCourses> getStudentCourseList() {
-		return repository.courseSearch();
-}
-
 }
