+++ src/main/java/raisetech/studentManagement/controller/StudentController.java
@@ -1,0 +1,30 @@
+package raisetech.studentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.studentManagement.data.Student;
+import raisetech.studentManagement.data.StudentsCourses;
+import raisetech.studentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentsCourses> getStudentCourseList() {
+    return service.searchStudentCourseList();
+  }
+}
