+++ src/main/java/raisetech/studentManagement/repository/StudentRepository.java
@@ -1,8 +1,10 @@
-package raisetech.studentManagement;
+package raisetech.studentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.studentManagement.data.Student;
+import raisetech.studentManagement.data.StudentsCourses;
 
 @Mapper
 public interface StudentRepository {
