+++ src/main/java/raisetech/studentManagement/service/StudentService.java
@@ -1,0 +1,29 @@
+package raisetech.studentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.studentManagement.data.Student;
+import raisetech.studentManagement.data.StudentsCourses;
+import raisetech.studentManagement.repository.StudentRepository;
+
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    //何かしらの処理を行うために、サービスを作成する。
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    return repository.search();
+  }
+
+  public List<StudentsCourses> searchStudentCourseList() {
+    return repository.courseSearch();
+  }
+}
